requirements.txt
dbt-core==1.7.4
dbt-duckdb==1.7.2
duckdb==0.10.0
apache-airflow==2.8.2
pandas==2.2.0
pyarrow==14.0.2
streamlit==1.31.0
plotly==5.18.0
great-expectations==0.18.8
numpy==1.26.3
faker==22.5.1
requests==2.31.0
tqdm==4.66.2
# Tests
pytest==8.0.0
pytest-cov==4.1.0
# DevOps / DataOps
dvc==3.42.0
pre-commit==3.6.2
ruff==0.3.2
black==24.2.0
isort==5.13.2
# Sécurité
detect-secrets==1.4.0
